removed need to use script task to set project.version property :
tstamp task is used to set the property, and in the main NAnt build file the overwrite attribute is set to false on the property task that normally sets the project.version property to prevent it from overwrite the value that's set in this build file.

that way we don't have to store the NAnt.Win32 assembly in cvs, as the script task is now located in the assembly.