always execute NAnt in new appdomain instead of the default domain, as AppDomain.AppendPrivatePath is deprecated in .NET 2.0. The reason for the deprecation is :
"The point of deprecating it is so that an appdomain's binding context can't be modified after assemblies have been loaded. Providing an alternative which allows that would defeat the purpose.

This is required for domain neutral assembly binding. Apps using this method should use AppDomainSetup.PrivateBinPath instead, at appdomain creation time. It was never intended that AppendPrivatePath() be used to append paths after assemblies have already been loaded. If that is desired, a new appdomain is recommended. "